Suggest creating a new lambda when a funcitonal interface is needed
- many fixes to parameter completion
- implementation note:

given the following code snippet:

```
myMethod(12, , );
```

there are broken ast nodes for the 2nd and 3rd argument.
I use these nodes to help build completion when completing the 2nd and 3rd argument,
since if the broken nodes didn't exist it would be really difficult to figure out
which argument is being completed.

Signed-off-by: David Thompson <davthomp@redhat.com>